Added weighting to item pedistool spawns and added some animation.
As above, added some weighting to the chosing of items. Also added proximity based U.I. and animations to the pedistools. Only needs implementation on the player end to use the consumables. The shop is complete from this point on barring re-skinning. Also cleaned up the project folder a bit, namely prefabs.